modal showing up on FinalTechChat
