+++ dart-cart-client/package.json
@@ -24,6 +24,7 @@
     "firebase": "^9.6.9",
     "mdb-react-ui-kit": "^2.3.0",
     "msw": "^0.38.1",
+    "net": "^1.0.2",
     "node-sass": "^7.0.1",
     "react": "^17.0.2",
     "react-bootstrap": "^2.2.0",
